Tests/front end/form page2

+++ db/migrations/testing/20190604202022_adjust_foreign_key_constraint.js
@@ -1,31 +1,39 @@
 exports.up = function(knex, Promise) {
   return Promise.all([
-    knex.schema.table("contacts_submissions", function(table) {
-      table.dropForeign("contact_id");
-      table.dropForeign("submission_id");
-    }),
-    knex.schema.table("contacts_submissions", function(table) {
-      table
-        .foreign("contact_id")
-        .references("contacts.contact_id")
-        .onDelete("CASCADE");
-      table
-        .foreign("submission_id")
-        .references("submissions.submission_id")
-        .onDelete("CASCADE");
+    knex.schema.hasTable("contacts_submissions").then(function(exists) {
+      if (exists) {
+        return knex.schema.table("contacts_submissions", function(table) {
+          table.dropForeign("contact_id");
+          table.dropForeign("submission_id");
+        });
+      }
+      return knex.schema.table("contacts_submissions", function(table) {
+        table
+          .foreign("contact_id")
+          .references("contacts.contact_id")
+          .onDelete("CASCADE");
+        table
+          .foreign("submission_id")
+          .references("submissions.submission_id")
+          .onDelete("CASCADE");
+      });
     })
   ]);
 };
 
 exports.down = function(knex, Promise) {
   return Promise.all([
-    knex.schema.table("contacts_submissions", function(table) {
-      table.dropForeign("contact_id");
-      table.dropForeign("submission_id");
-    }),
-    knex.schema.table("contacts_submissions", function(table) {
-      table.foreign("contact_id").references("contacts.contact_id");
-      table.foreign("submission_id").references("submissions.submission_id");
+    knex.schema.hasTable("contacts_submissions").then(function(exists) {
+      if (exists) {
+        return knex.schema.table("contacts_submissions", function(table) {
+          table.dropForeign("contact_id");
+          table.dropForeign("submission_id");
+        });
+      }
+      return knex.schema.table("contacts_submissions", function(table) {
+        table.foreign("contact_id").references("contacts.contact_id");
+        table.foreign("submission_id").references("submissions.submission_id");
+      });
     })
   ]);
 };
